Created api to find politiebureaus.
Created two endpoints. One with a simple look-up which becomes less accurate the bigger the distance.

Another one which has faster lookup due to the usage of a Quadtree and more accurate due to the use of the Haversine formula.

Unit tests included.

pom.xml were for some reason ignored.